detector_service/requirements.txt
flask
requests
torch
joblib
scikit-learn
numpy
gunicorn
psycopg2-binary
